requirements.txt
langchain==0.0.340
langchain-core==0.1.0
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
